pyproject.toml
[project]
name = "heatmap-globular-adsb"
version = "0.1.0"
description = "Scheduled ADS-B flight fetcher and heatmap generator for Globular ADS-B"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flightradarapi",
    "numpy",
    "Pillow",
    "scipy",
    "boto3",
    "schedule",
    "beautifulsoup4>=4.14.3",
]

[project.scripts]
globular-pipeline  = "globular_adsb.pipeline:main"
globular-scheduler = "globular_adsb.scheduler:main"
globular-dev       = "globular_adsb.dev_server:main"
globular-generate-heatmap = "globular_adsb.generate:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "moto[s3]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/globular_adsb"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/globular_adsb"]
